Desktop/tiktok/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-generativeai>=0.8.0
google-api-python-client>=2.111.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
